Big re-org of the agents and estimators
Getting my head around what agetns and estimators we have.
